app/requirements.txt
Flask==3.0.0
flask_sqlalchemy==3.1.1
pandas==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
